Region alloc tracking: Fix bug where the iteration fails to stop at zero, causing a segfault

+++ src/misc/regionalloctracking.cpp
@@ -47,7 +47,7 @@ Bitu RegionAllocTracking::getMemory(Bitu bytes,const char *who,Bitu alignment,Bi
 	{
 		/* allocate downward from the top */
 		si = topDownAlloc ? (alist.size() - 1) : 0;
-		while (si >= 0) {
+		while ((ssize_t)si >= 0) {
 			Block &blk = alist[si];
 
 			if (!blk.free || (blk.end+1-blk.start) < bytes) {
